test(github): share one Check Run capture struct across webhook tests
Three near-identical structs decoded the fake GitHub server's Check Run
payloads, and one lived in an untagged file that an integration-tagged
test silently depended on. A single shared untagged definition makes
that dependency explicit and gives every test the same field set.
