+++ pkg/webhook/durable_push_test.go
@@ -100,9 +100,9 @@ func TestDurablePushWebhookIgnoresFeatureBranch(t *testing.T) {
 // each gated environment, then marks the delivery completed.
 func TestDurablePushDriverPostsChecks(t *testing.T) {
 	client, mux := setupGitHubServer(t)
-	created := make(chan createdCheckRun, 10)
+	created := make(chan checkRunCapture, 10)
 	mux.HandleFunc("POST /repos/octocat/hello-world/check-runs", func(w http.ResponseWriter, r *http.Request) {
-		var c createdCheckRun
+		var c checkRunCapture
 		require.NoError(t, json.NewDecoder(r.Body).Decode(&c))
 		created <- c
 		w.WriteHeader(http.StatusCreated)
